+++ src/Handlers/CorsPreflightHandler.php
@@ -9,16 +9,30 @@ class CorsPreflightHandler extends BaseHandler
 {
     private $response;
 
-    public function __construct(Response $response)
-    {
+    private $headers = [];
+
+    public function __construct(
+        Response $response,
+        $headers = [
+            'Access-Control-Allow-Headers' => [
+                'Authorization',
+                'X-Requested-With',
+            ],
+        ]
+    ) {
         parent::__construct();
         $this->response = $response;
+        $this->headers = $headers;
     }
 
     public function handle($params)
     {
-        $this->response->addHeader('Access-Control-Allow-Headers', 'Authorization');
-        $this->response->addHeader('Access-Control-Allow-Headers', 'X-Requested-With');
+        foreach ($this->headers as $name => $values) {
+            $values = is_array($values) ? $values : [$values];
+            foreach ($values as $value) {
+                $this->response->addHeader($name, $value);
+            }
+        }
         return new JsonApiResponse(Response::S200_OK, []);
     }
 }